+++ public_api.ts
@@ -1,7 +1,7 @@
 /**
  * Angular library starter.
- * Build an Angular library compatible with AoT compilation & Tree shaking.
- * Written by Roberto Simonetti.
+ * Build an Angular library compatible with AoT compilation & Tree shaking
+ * Copyright Roberto Simonetti.
  * MIT license.
  * https://github.com/robisim74/angular-library-starter
  */
@@ -10,4 +10,3 @@
  * Entry point for all public APIs of the package.
  */
 export { NgxTypeaheadModule } from './src/modules/ngx-typeahead.module';
-