+++ src/modules/ngx-typeahead.module.ts
@@ -1,13 +1,22 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HttpModule, JsonpModule } from '@angular/http';
+import { HttpClientModule, HttpClientJsonpModule } from '@angular/common/http';
+
+import 'rxjs/add/operator/debounceTime';
+import 'rxjs/add/operator/concat';
+import 'rxjs/add/operator/distinctUntilChanged';
+import 'rxjs/add/operator/share';
+import 'rxjs/add/operator/filter';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/switchMap';
+import 'rxjs/add/observable/fromEvent';
 
 import { NgxTypeAheadComponent } from './ngx-typeahead.component';
 
 @NgModule({
   declarations: [NgxTypeAheadComponent],
   exports: [NgxTypeAheadComponent, CommonModule],
-  imports: [CommonModule, HttpModule, JsonpModule],
+  imports: [CommonModule, HttpClientModule, HttpClientJsonpModule],
   providers: []
 })
-export class NgxTypeaheadModule {}
+export class NgxTypeaheadModule { }