+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v 0.0.4 (2017/11/05)
 * Replaced deprecated ngOutletContext by ngTemplateOutletContext to support Angular version 5.
+* Added taItemTpl example to Readme
 
 ## v 0.0.3 (2017/08/28)
 * Added Api for choosing protocols (jsonp or http)